requirements.txt
# Core ML
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.3.0
torch==2.0.1
transformers==4.35.0

# ML Models
lightgbm==4.0.0
xgboost==1.7.6
prophet==1.1.4
implicit==0.7.0

# Data Processing
scipy==1.11.0
networkx==3.1
pyarrow==12.0.0

# GCP
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.0
google-cloud-aiplatform==1.35.0

# API
fastapi==0.103.0
uvicorn==0.23.0
pydantic==2.4.0

# Utils
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.0
joblib==1.3.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0